Refactor code for loading Steam games
